Update README.md
The links to the demo applications need to be https, not http. At least, this is true for Chrome. This was a security upgrade change Chrome mandated. If the objects are not created under https then you get an alert that says "hacksparrow.github.io says: object NavigatorUserMediaError". Navigating to https://hacksparrow.github.io/remote.js/example/controlled.html fixes the problem.